+++ src/lib/lib.rs
@@ -3,6 +3,7 @@ mod error;
 mod ops;
 mod pr_title;
 mod utilities;
+mod workspace;
 
 pub use client::Client;
 pub use error::Error;
@@ -12,3 +13,4 @@ pub use ops::MakeRelease;
 pub use ops::Sign;
 pub use ops::UpdateFromPr;
 pub use pr_title::PrTitle;
+pub use workspace::{Package, Workspace};